Datos normativos afectivos de emojis

Valores normativos afectivos para un conjunto de emojis, recogidos en poblacion argentina. Cada emoji aparece en tres filas, una por submuestra normativa: poblacion general, mujeres y hombres.

Usage

emojis

Format

Un data frame con 180 filas y 12 variables:

emoji

Caracter Unicode del emoji.

poblacion

Submuestra normativa: general, Mujeres u Hombres.

val_media

Media de valencia (agrado).

val_desvio

Desvio estandar de valencia.

aro_media

Media de activacion (arousal).

aro_desvio

Desvio estandar de activacion.

dom_media

Media de dominancia.

dom_desvio

Desvio estandar de dominancia.

fam_media

Media de familiaridad.

fam_desvio

Desvio estandar de familiaridad.

cla_media

Media de claridad.

cla_desvio

Desvio estandar de claridad.

Source

Datos normativos propios. Proyecto OSF: https://osf.io/p75xr/

Details

Las dimensiones afectivas (valencia, activacion, dominancia) siguen el modelo dimensional del afecto. Familiaridad y claridad son indices adicionales de procesamiento del estimulo. Todas las variables se midieron en escalas tipo Likert de 1 a 9.

Examples

data(emojis)
head(emojis)
#>   emoji poblacion val_media val_desvio aro_media aro_desvio dom_media
#> 1    😠   general      3.60       2.13      5.81       2.28      4.89
#> 2    😧   general      3.53       2.18      5.32       2.07      4.19
#> 3    😰   general      3.47       2.21      5.12       2.06      3.99
#> 4    😲   general      5.39       1.66      5.08       2.11      5.18
#> 5    😖   general      3.51       2.04      5.21       2.06      4.41
#> 6    😕   general      4.15       1.48      4.34       2.04      4.86
#>   dom_desvio fam_media fam_desvio cla_media cla_desvio
#> 1       2.22      3.82       2.03      6.43       1.17
#> 2       2.05      2.84       1.84      5.02       1.81
#> 3       2.02      3.07       1.76      4.78       1.83
#> 4       2.04      3.42       2.01      5.47       1.81
#> 5       2.04      3.26       1.93      4.57       1.96
#> 6       2.00      4.07       2.08      5.20       1.80
# Solo la norma de poblacion general
subset(emojis, poblacion == "general")
#>    emoji poblacion val_media val_desvio aro_media aro_desvio dom_media
#> 1     😠   general      3.60       2.13      5.81       2.28      4.89
#> 2     😧   general      3.53       2.18      5.32       2.07      4.19
#> 3     😰   general      3.47       2.21      5.12       2.06      3.99
#> 4     😲   general      5.39       1.66      5.08       2.11      5.18
#> 5     😖   general      3.51       2.04      5.21       2.06      4.41
#> 6     😕   general      4.15       1.48      4.34       2.04      4.86
#> 7     😓   general      3.41       1.98      4.54       2.20      4.46
#> 8     🤤   general      6.44       2.38      4.93       2.43      5.92
#> 9     🤯   general      5.45       2.07      6.19       2.50      5.10
#> 10    😑   general      4.81       1.38      4.23       2.10      5.21
#> 11    😘   general      6.77       2.38      4.69       2.56      6.34
#> 12    😱   general      4.99       2.00      5.89       2.38      4.62
#> 13    🤮   general      2.91       2.52      6.05       2.47      3.60
#> 14    🤭   general      6.40       2.44      4.78       2.41      5.59
#> 15    🤕   general      3.59       2.35      4.99       2.18      3.86
#> 16    🧐   general      5.14       1.37      4.49       2.10      5.89
#> 17    😶   general      4.63       1.66      4.45       2.41      4.70
#> 18     🫣   general      5.20       1.73      5.37       2.27      4.87
#> 19    🤨   general      4.85       1.54      4.96       1.98      5.21
#> 20    🙄   general      4.25       1.68      5.36       2.16      4.98
#> 21    😤   general      3.42       2.14      6.49       2.35      5.06
#> 22    🤬   general      3.42       2.40      6.76       2.56      5.09
#> 23    😂   general      7.31       2.66      5.25       2.83      6.87
#> 24    😛   general      6.63       2.14      4.12       2.37      6.37
#> 25    😨   general      3.75       2.10      5.91       2.31      3.88
#> 26    😳   general      5.02       1.69      5.62       2.26      4.29
#> 27    🙁   general      3.06       2.51      5.03       2.16      3.77
#> 28    😬   general      4.52       1.52      4.88       1.95      4.71
#> 29    😀   general      7.18       2.44      4.58       2.63      6.68
#> 30    😅   general      6.17       2.08      5.06       2.14      5.74
#> 31    😆   general      7.17       2.41      5.20       2.55      6.58
#> 32    🤗   general      6.97       2.40      4.34       2.69      6.65
#> 33    😭   general      2.78       2.66      6.09       2.64      3.23
#> 34    🤥   general      4.78       1.99      4.74       2.29      4.86
#> 35     🫠   general      4.88       1.85      4.75       2.23      4.38
#> 36    😐   general      4.84       1.23      4.11       2.09      5.17
#> 37    🥳   general      7.38       2.58      5.69       2.71      6.81
#> 38    😔   general      3.11       2.43      5.22       2.24      3.53
#> 39    🥺   general      4.48       2.72      5.26       2.35      4.33
#> 40    😡   general      3.07       2.55      6.97       2.64      4.31
#> 41    🤫   general      5.08       1.41      4.60       2.21      5.55
#> 42    😴   general      5.56       2.20      3.06       2.60      5.90
#> 43    🙂   general      5.84       1.87      3.95       2.32      5.78
#> 44    😊   general      6.89       2.16      4.07       2.44      6.35
#> 45    😇   general      6.57       2.06      3.39       2.49      6.33
#> 46    😍   general      7.24       2.63      5.46       2.92      5.61
#> 47    🥰   general      7.34       2.55      5.33       2.80      5.89
#> 48    😄   general      6.94       2.21      4.27       2.44      6.29
#> 49    😎   general      6.92       2.41      4.24       2.59      7.21
#> 50    🥲   general      4.23       2.10      4.93       1.96      4.24
#> 51    😏   general      6.29       2.10      5.28       2.32      6.06
#> 52    🤩   general      7.22       2.53      5.24       2.59      6.65
#> 53    🤔   general      5.09       1.23      4.77       1.96      5.11
#> 54    😒   general      3.92       1.90      4.96       2.09      4.49
#> 55    😩   general      3.43       2.41      5.63       2.21      3.84
#> 56    😉   general      6.49       1.97      4.31       2.21      6.37
#> 57    😟   general      3.19       2.22      5.10       2.18      3.65
#> 58    🥱   general      5.23       1.81      3.88       2.57      4.93
#> 59    🤪   general      6.67       2.33      5.29       2.30      6.03
#> 60    🤐   general      4.14       1.82      5.12       2.11      4.41
#>    dom_desvio fam_media fam_desvio cla_media cla_desvio
#> 1        2.22      3.82       2.03      6.43       1.17
#> 2        2.05      2.84       1.84      5.02       1.81
#> 3        2.02      3.07       1.76      4.78       1.83
#> 4        2.04      3.42       2.01      5.47       1.81
#> 5        2.04      3.26       1.93      4.57       1.96
#> 6        2.00      4.07       2.08      5.20       1.80
#> 7        2.15      3.71       1.95      4.70       1.71
#> 8        2.26      4.68       2.09      5.84       1.62
#> 9        2.27      3.68       2.10      5.66       1.64
#> 10       2.00      4.11       2.11      4.89       1.95
#> 11       2.29      5.73       1.75      6.38       1.30
#> 12       2.26      4.59       2.05      6.14       1.46
#> 13       2.09      3.77       2.18      6.44       1.31
#> 14       2.29      4.60       1.91      5.27       1.67
#> 15       2.22      2.53       1.86      4.83       1.92
#> 16       2.27      3.13       2.03      4.03       2.00
#> 17       2.34      3.20       2.02      3.99       2.10
#> 18       2.14      3.22       2.00      4.55       1.91
#> 19       2.07      4.88       1.94      5.70       1.54
#> 20       2.27      5.33       1.69      5.99       1.40
#> 21       2.54      3.91       2.08      5.75       1.53
#> 22       2.78      3.87       2.30      6.56       1.14
#> 23       2.20      5.93       1.61      6.69       0.89
#> 24       2.32      3.35       1.91      3.75       1.94
#> 25       2.24      3.15       1.89      5.03       1.61
#> 26       2.10      4.56       1.93      5.43       1.62
#> 27       2.14      4.20       2.06      5.95       1.44
#> 28       1.94      4.12       2.14      4.42       2.09
#> 29       2.16      4.95       2.07      6.26       1.36
#> 30       2.12      4.31       1.99      5.04       1.75
#> 31       2.19      3.91       2.25      5.19       1.84
#> 32       2.25      4.58       2.07      4.68       1.88
#> 33       2.37      5.62       1.82      6.62       1.22
#> 34       2.28      2.32       1.72      4.86       2.16
#> 35       2.03      3.78       2.26      3.69       2.14
#> 36       1.91      4.53       1.93      5.13       1.75
#> 37       2.15      4.96       2.08      6.57       1.02
#> 38       2.00      5.21       1.79      6.22       1.22
#> 39       2.39      5.37       1.90      5.60       1.61
#> 40       2.90      4.57       2.22      6.71       1.01
#> 41       2.36      3.67       2.12      5.91       1.59
#> 42       2.57      5.29       2.01      6.70       1.02
#> 43       2.12      4.54       1.99      5.10       1.99
#> 44       2.04      4.93       1.87      5.71       1.48
#> 45       2.20      3.42       1.98      5.12       1.85
#> 46       2.63      5.69       1.82      6.68       0.93
#> 47       2.43      6.15       1.41      6.50       1.08
#> 48       2.08      5.10       1.73      5.73       1.48
#> 49       2.18      4.81       1.97      5.22       1.66
#> 50       2.01      4.25       2.03      5.10       1.68
#> 51       2.13      4.58       2.02      5.54       1.64
#> 52       2.12      3.54       2.24      4.02       2.03
#> 53       1.93      5.21       1.80      5.97       1.45
#> 54       2.05      4.72       1.99      5.80       1.43
#> 55       1.99      4.42       2.07      5.32       1.74
#> 56       2.04      4.36       2.07      5.42       1.68
#> 57       1.85      3.47       1.93      5.18       1.70
#> 58       2.15      4.49       2.21      5.91       1.57
#> 59       1.98      3.54       2.16      3.83       1.96
#> 60       2.38      3.15       2.14      5.22       1.95